Handle delete caategory and Tag

+++ app/Http/Controllers/CategoryController.php
@@ -91,6 +91,10 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        if ($category->posts->count() > 0) {
+            session()->flash('error', 'Category can not be Deleted because it has some posts');
+            return redirect()->back();
+        }
         $category->delete();
 
         session()->flash('success', 'Category Deleted Successfully');
