+++ app/Http/Controllers/TagsController.php
@@ -89,6 +89,10 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        if ($tag->posts->count() > 0) {
+            session()->flash('error', 'Tag can not be Deleted because it has associated with some posts');
+            return redirect()->back();
+        }
         $tag->delete();
 
         session()->flash('success', 'Tag Deleted Successfully');
